% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-data.R
\docType{data}
\name{game_2x2_structures}
\alias{game_2x2_structures}
\title{Structure of 2x2 games from Bruns, 2015}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 7 columns.
}
\usage{
game_2x2_structures
}
\description{
Structure of 2x2 games from Bruns, 2015
}
\keyword{internal}
