% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_2x2_contamination_vector}
\alias{get_2x2_contamination_vector}
\title{Get 2x2 contamination vector}
\usage{
get_2x2_contamination_vector(type, player)
}
\arguments{
\item{type}{Player type as 1, 2, 3, or 4}

\item{player}{Player A or player B}
}
\description{
Get 2x2 contamination vector
}
\details{
The player types are:
\enumerate{
\item Contaminate no one
\item Contaminate self
\item Contaminate other player
\item Contaminate both
}
Contamination structure C_A, C_B: vector indicating cost to each player
\itemize{
\item C_A1: c(0,0)
\item C_A2: c(1,0)
\item C_A3: c(0,1)
\item C_A2: c(1,1)
\item C_B1: c(0,0)
\item C_B2: c(0,1)
\item C_B3: c(1,0)
\item C_B2: c(1,1)
}
}
\examples{
payouts <- get_2x2_payouts(3, 3, Cs = 1, Cd = 2, T)
get_2x2_ggplot(payouts)

payouts <- get_2x2_payouts(3, 3, Cs = 2, Cd = 1, T)
get_2x2_ggplot(payouts)

payouts <- get_2x2_payouts(4, 4, Cs = 1, Cd = 2, T)
get_2x2_ggplot(payouts)
}
\keyword{internal}
