% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_2x2_ggplot}
\alias{get_2x2_ggplot}
\title{ggplot Payouts}
\usage{
get_2x2_ggplot(payouts, equilibria = FALSE, nsmall = 2)
}
\arguments{
\item{payouts}{Tibble of payouts from get_2x2_payouts}

\item{equilibria}{If \code{TRUE}, highlight equilibria}

\item{nsmall}{Used for rounding payouts}
}
\description{
Generate ggplot of 2x2 payouts
}
\examples{
payouts <- get_2x2_payouts(3, 3, Cs = 1, Cd = 2, TRUE)
get_2x2_ggplot(payouts)

payouts <- get_2x2_payouts(3, 3, Cs = 2, Cd = 1, TRUE)
get_2x2_ggplot(payouts, TRUE)

payouts <- get_2x2_payouts(4, 4, Cs = 1, Cd = 2, TRUE)
get_2x2_ggplot(payouts, TRUE)
}
