% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_2x2_nash_stability}
\alias{get_2x2_nash_stability}
\title{Get 2x2 Nash stability}
\usage{
get_2x2_nash_stability(payouts, i)
}
\arguments{
\item{payouts}{List of payouts from get_payouts_2x2}

\item{i}{Index at which to evaluate Nash stability}
}
\value{
Returns a value of 1 (Nash equilibrium), 0 (stable), or -1 (not NE)
}
\description{
This function evaluates Nash stability by checking if either player
would benefit by switching their action.
}
\details{
Get 2x2 Nash equilibrium and first best
}
\examples{
payouts <- get_payouts_2x2(3, 3, Cs = 2, Cd = 3, T)
payouts <- get_payouts_2x2(3, 3, Cs = 2, Cd = 3)
get_2x2_nash_stability(payouts, 4)
get_2x2_nash_stability(payouts, 1)
}
