% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_2x2_weighted_payouts}
\alias{get_2x2_weighted_payouts}
\title{Get 2x2 weighted payouts}
\usage{
get_2x2_weighted_payouts(..., weights)
}
\arguments{
\item{...}{Tibbles of payouts from get_payouts_2x2}

\item{weights}{Vector used to weight payouts, in order of \code{...}}
}
\description{
Get 2x2 weighted payouts
}
\examples{
payouts_i <- get_payouts_2x2(3, 3, Cs = 2, Cd = 3, T)
payouts_ii <- get_payouts_2x2(3, 1, Cs = 2, Cd = 3, T)

weights <- c(0.5, 0.5)
weighted_payouts <- get_2x2_weighted_payouts(payouts_i, payouts_ii, weights = weights)
get_2x2_ggplot(weighted_payouts)
}
