% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-model.R
\name{get_contamination_probabilities}
\alias{get_contamination_probabilities}
\title{Get probabilities of contamination}
\usage{
get_contamination_probabilities(params_df, show_progress = FALSE)
}
\arguments{
\item{params_df}{\code{tibble} with appropriate columns (see details)}

\item{show_progress}{boolean, whether or not to show progress bar}
}
\description{
Get probabilities of contamination
}
\details{
\code{params_df} must contain the following named columns:
\itemize{
\item \code{z1}, \code{z2},  \code{rs}, \code{density}, \code{area} must be \code{units} objects.
\item \code{self_treat} is a boolean
\item \code{theta_range} should be in radians as \code{c(theta_min, theta_max)}
\item \code{alpha_range} should be in radians as \code{c(alpha_min, alpha_max)}
\item \code{hh_array_type} character as "septic" or "well" (see \code{get_septic_well_array})
}
}
\examples{
library(units)
library(tidyr)
params_row <- tibble(
  z1 = set_units(10, "ft"),
  z2 = set_units(20, "ft"),
  area = set_units(64, "acre"),
  theta_range = list(c(0, pi/4)), # this will be unlisted in the function
  alpha_range = list(c(0, 20)), # this will be unlisted in the function
  hh_array_type = "septic")

# This function allows multiple rows to be calculated at once
params_df <- params_row \%>\%
  crossing(density = set_units(c(0, 0.5), "1/acre"),
  rs = set_units(c(10, 20),"ft"), self_treat = c(TRUE, FALSE))

params_df$probs <- get_contamination_probabilities(params_df)
params_df
}
