% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-model.R
\name{get_hh_grid}
\alias{get_hh_grid}
\title{Get household grid}
\usage{
get_hh_grid(density, area)
}
\arguments{
\item{density}{housing density as a units object with units [1/area]}

\item{area}{total area of area as an object with units [acre]}
}
\description{
Get household grid
}
\details{
Create a grid of septic fields around a well at (x = 0, y = 0). If the grid
length is even-numbered, the final row and column are place at positive x and y.
}
\examples{
library(units)
density <- set_units(0.25,"1/acre")
area <- set_units(64,"acre")
hh_grid <- get_hh_grid(density,area)

library(ggplot2)
library(ggforce) # ggforce allows units objects for the axes
ggplot(hh_grid) +
  geom_point(aes(x, y)) +
  geom_point(aes(0, 0), color = "red", size = 3, shape = 1)
}
