% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-model.R
\name{get_intersection_probability}
\alias{get_intersection_probability}
\title{Get intersection probability}
\usage{
get_intersection_probability(
  wells_array,
  theta_range = c(-pi, pi),
  alpha_range = c(0, 100),
  self_treat = FALSE,
  return_option = 1,
  show_progress = TRUE
)
}
\arguments{
\item{wells_array}{Wells object prepared with \code{get_septic_well_array}}

\item{theta_range}{Vector describing the min and max of the uniform distribution for the mean lateral direction of flow}

\item{alpha_range}{Vector describing the min and max of distance to the groundwater divide}

\item{self_treat}{Logical. If \code{TRUE}, any well at (x = 0, y = 0) will be removed}

\item{return_option}{Either 1 or 2. See Return, below}

\item{show_progress}{Logical that determines if progress bar is shown}
}
\value{
If return_option = 1: Probability of contamination
If return_option = 2: Returns a list containing the well array after clipping and shifting and probabilities of contamination
}
\description{
Get intersection probability
}
\details{
Note that theta_range must not cross -pi or pi.
}
\examples{
library(units)
wells_array <- get_septic_well_array(hh_grid_example, "septic")
prob <- get_intersection_probability(wells_array, show_progress = FALSE)
}
