% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_payouts_2x2}
\alias{get_payouts_2x2}
\title{Get payouts for 2x2 games}
\usage{
get_payouts_2x2(tB, tA, Cs, Cd, pos = FALSE)
}
\value{
The function returns a tibble containing the following columns:
\itemize{
\item A, B: player decision as either 1 (Upgrade) or 0 (Neglect)
\item tA, tB: Type of each player
\item Cs_i: Expected utility (i.e., cost) of upgrading septic system
\item Cd_i: Expected utility (i.e., cost) of contaminated domestic well
\item UA, UB: Expected utility of both costs
\item payouts: Payout structure for each cell for A, B
}
}
\description{
Get payouts for 2x2 games
}
