% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-model.R
\name{get_row_contamination_probability}
\alias{get_row_contamination_probability}
\title{Get probability of contamination of row}
\usage{
get_row_contamination_probability(params_row)
}
\arguments{
\item{params_row}{\code{list} or \code{tibble} row with appropriate names variables (see details)}
}
\description{
Get probability of contamination of row
}
\details{
\code{params_row} must contain the following named variables:
\itemize{
\item \code{z1}, \code{z2},  \code{rs}, \code{density}, \code{area} must be \code{units} objects.
\item \code{self_treat} is a boolean
\item \code{theta_range} should be in radians as \code{c(theta_min, theta_max)}
\item \code{alpha_range} should be in radians as \code{c(alpha_min, alpha_max)}
}
}
\examples{
\dontrun{
library(units)
params_row <- tibble(
  z1 = set_units(10, "ft"),
  z2 = set_units(20, "ft"),
  rs = set_units(15, "ft"),
  density = set_units(1, "1/acre"),
  area = set_units(64, "acre"),
  self_treat = TRUE,
  theta_range = list(c(0, pi/4)), # this will be unlisted in the function
  alpha_range = list(c(0, 20))) # this will be unlisted in the function
get_row_contamination_probability(params_row)

# This function allows multiple rows to be calculated at once
params_df <- params_row \%>\% dplyr::select(-density, -rs, -self_treat) \%>\%
  tidyr::crossing(density = set_units(c(0, 1), "1/acre"), rs = set_units(c(10, 20),"ft"), self_treat = c(TRUE, FALSE))

# use sapply to get all probabilities at once
params_df$probs <- sapply(split(params_df, 1:nrow(params_df)), get_row_contamination_probability)
params_df
}
}
\keyword{internal}
