% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-model.R
\name{get_union_probability}
\alias{get_union_probability}
\title{Get probability of contamination}
\usage{
get_union_probability(
  wells_array,
  theta_range = c(-pi, pi),
  alpha_range = c(0, 100),
  show_progress = TRUE
)
}
\arguments{
\item{wells_array}{Wells object prepared with \code{get_septic_well_array}}

\item{theta_range}{Vector describing the min and max of the uniform distribution for the mean lateral direction of flow}

\item{alpha_range}{Vector describing the min and max of distance to the groundwater divide}

\item{show_progress}{Logical that determines if progress bar is shown}
}
\description{
Get probability of contamination
}
\details{
Careful: this won't work if theta_range cross over -pi or pi
}
\keyword{internal}
