% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-model.R
\name{get_well_rectangle}
\alias{get_well_rectangle}
\title{Get rectangle polygon from coordinates}
\usage{
get_well_rectangle(theta1, theta2, phi1, phi2)
}
\arguments{
\item{theta1}{Lower limit theta}

\item{theta2}{Upper limit theta}

\item{phi1}{Lower limit phi}

\item{phi2}{Upper limit phi}
}
\description{
Get rectangle polygon in theta-phi space from coordinates
}
\details{
Returns a rectangle as an sf polygon in theta-phi space
}
\examples{
\dontrun{
get_well_rectangle(0,pi/2,0,pi/6)
}
}
\keyword{internal}
