% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-data.R
\docType{data}
\name{hh_grid_example}
\alias{hh_grid_example}
\title{Gridded household data}
\format{
An object of class \code{data.frame} with 16 rows and 8 columns.
}
\usage{
hh_grid_example
}
\description{
Example data to get started containing a 4x4 grid of households
}
\keyword{datasets}
