% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw-model.R
\name{shift_hh_grid_pi}
\alias{shift_hh_grid_pi}
\title{Shift rectangle polygons to fall within [-pi, pi]}
\usage{
shift_hh_grid_pi(wells_array, theta_range = c(-pi, pi))
}
\arguments{
\item{wells_array}{Wells object prepared with \code{get_septic_well_array}}

\item{theta_range}{Vector describing the min and max of the uniform distribution for the mean lateral direction of flow}
}
\description{
Shift rectangle polygons to fall within [-pi, pi]
}
\details{
Shift all well rectangles by ±2pi and clip to the range \code{[theta_min, theta_max]}. These theta values
don't necessarily need to be the extremes [-pi, pi] and can be any subset of that range.
}
\keyword{internal}
