% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_2x2_game_solutions}
\alias{get_2x2_game_solutions}
\title{Get 2x2 game solutions}
\usage{
get_2x2_game_solutions(payouts)
}
\arguments{
\item{payouts}{List of payouts from get_2x2_payouts}
}
\description{
This function evaluates Nash stability by checking if
}
\details{
Get 2x2 Nash equilibrium and first best
}
\examples{
payouts <- get_2x2_payouts(3, 3, Cs = 2, Cd = 3)
get_2x2_game_solutions(payouts)
get_2x2_ggplot(payouts, equilibria = TRUE)
payouts_ii <- get_2x2_payouts(3, 1, Cs = 2, Cd = 3, T)

weights <- c(0.5, 0.5)
weighted_payouts <- get_2x2_weighted_payouts(list(payouts_i, payouts_ii), weights = weights)
get_2x2_ggplot(weighted_payouts)
}
