% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_2x2_ggplot}
\alias{get_2x2_ggplot}
\title{ggplot Payouts}
\usage{
get_2x2_ggplot(payouts)
}
\arguments{
\item{payouts}{Tibble of payouts from get_payouts_2x2}
}
\description{
Generate ggplot of 2x2 payouts
}
