% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games-2x2.R
\name{get_2x2_weighted_payouts}
\alias{get_2x2_weighted_payouts}
\title{Get 2x2 weighted payouts}
\usage{
get_2x2_weighted_payouts(payouts_list, weights)
}
\arguments{
\item{payouts_list}{List of payouts from get_2x2_payouts}

\item{weights}{Vector used to weight payouts, in order of \code{...}}
}
\description{
Get 2x2 weighted payouts
}
\examples{
payouts_i <- get_2x2_payouts(3, 3, Cs = 2, Cd = 3, TRUE)
payouts_ii <- get_2x2_payouts(3, 1, Cs = 2, Cd = 3, TRUE)

weights <- c(0.5, 0.5)
weighted_payouts <- get_2x2_weighted_payouts(list(payouts_i, payouts_ii), weights = weights)

library(gridExtra)
grid.arrange(get_2x2_ggplot(payouts_i, equilibria = TRUE),
  get_2x2_ggplot(payouts_ii, equilibria = TRUE),
  get_2x2_ggplot(weighted_payouts, equilibria = TRUE))
}
